Support folder tree navigation in cloud import
Only load a single folder at a time as the full tree traversal was
incredibly slow for large amounts of files
When a folder is selected its contents will be loaded and re-populate
the select dropdown, previous level [..] can be selected to go up the
tree.
We no longer load the image count in advance, it is only loaded on
opening the folder to further reduce slow webdav calls.