Use Uppy component for resumable uploads to tusd server
Replaces the existing Dropzone component to support resumable uploads
with future support for various cloud services.
Files are uploaded to a tusd server and added to the task as urls
instead of paths, when the task processing begins they are downloaded to
the local filesystem.
Known issues: adding large numbers of files (eg: > 500) is slow to
generate preview thumbnails, will need to disable previews in such
cases. Workaround is to add files in batches or just wait for the
processing to finish.